- overview page restructure
