+++ src/routes/BondRoutes.tsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import { Route, Switch, RouteComponentProps } from 'react-router-dom'
 
-import { Overview } from '../common/components/Bonds/OverviewWrapper/Overview'
+import { Overview } from '../pages/bond/overview'
 import Exchange from '../common/components/Bonds/ExchangeWrapper/Exchange'
 import Orders from '../modules/BondAccountOrders/BondAccountOrders.container'
